refactor(tui): split render.rs into a render/ module
render.rs was 1,690 lines holding seven unrelated subjects, so anything
touching the entries table sat next to onboarding copy. It divided on
subject with no cross-talk:

  render/mod.rs        300  ui, LayoutRow, LayoutRows
  render/overlay.rs     93  render_overlay, overlay_height, overlay_hints,
                            wrap, CURSOR_MARKER
  render/popups.rs     324  help, detail, confirm, HELP_SECTIONS
  render/onboarding.rs 186  both step screens, surface_blurb, step_dots
  render/surfaces.rs   261  marks, summary and pane surfaces
  render/entries.rs    346  entries table, overview, weekly breakdown
  render/form.rs       229  entry form, search bar

overlay.rs is the piece that earns the split on its own: it is the reusable
part every modal is built from, and it was buried among the things using it.

A pure move. Verified mechanically: normalising both sides (imports and
blank lines dropped, visibility prefixes stripped) gives 1,559 identical
body lines on each side, with exactly one intended difference —
render_help_popup narrows from `pub` to `pub(super)`, since `ui` is the only
thing outside this module that anything calls.

Visibility is otherwise as narrow as it goes: 18 items are `pub(super)`,
reachable from render/mod.rs and nowhere else; entry_row, day_header_row,
render_pane, HELP_SECTIONS, surface_blurb, step_dots and the two onboarding
step screens stay private to their file.

All 320 tests pass unchanged, 17 of which render real frames through this
code. fmt and clippy -D warnings clean.

Closes #54
